Adding documentation and breaking up large files
